PP-15145: Add method to return capture url to AdyenRequestUtil, refactor getHeaders method and create Adyen handler folder.
